# server/internal/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Latch Admin API
  version: 1.0.0
  description: Go + Gin + Casbin admin backend
servers:
  - url: /api/v1
paths:
...
  /users:
    get:
      summary: List users (paginated, data-scoped)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string }, description: Search username, nickname, email, phone, department, title, remark, last login IP }
        - { name: status, in: query, schema: { type: string } }
        - { name: gender, in: query, schema: { type: string } }
        - { name: department, in: query, schema: { type: string } }
        - { name: roleId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeUserPage"
  /logs:
    get:
      summary: List operation logs
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: username, in: query, schema: { type: string } }
        - { name: module, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeOpLogPage"
...
  /departments:
    get:
      summary: Department tree (paginated roots)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeDepartmentPage"
  /roles:
    get:
      summary: List roles
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeRolePage"
  /permissions:
    get:
      summary: List permissions
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
        - { name: kind, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopePermissionPage"
  /dicts:
    get:
      summary: List dictionary types
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeDictPage"
  /configs:
    get:
      summary: List system configs
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
        - { name: group, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeConfigPage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EnvelopeBase:
      type: object
      required: [code, message]
      properties:
        code: { type: integer }
        message: { type: string }
    User:
      type: object
      required: [id, username, nickname, avatar, email, phone, gender, department, title, remark, status, timezone, marketingOptIn, roles]
      properties:
        id: { type: integer }
        username: { type: string }
        nickname: { type: string }
        avatar: { type: string }
        email: { type: string }
        phone: { type: string }
        gender: { type: string }
        department: { type: string }
        title: { type: string }
        remark: { type: string }
        status: { type: string }
        timezone: { type: string }
        marketingOptIn: { type: boolean }
        lastLoginAt: { type: string, nullable: true }
        lastLoginIp: { type: string }
        permissionCodes: { type: array, items: { type: string } }
        roles: { type: array, items: { $ref: "#/components/schemas/Role" } }
        createdAt: { type: string }
        updatedAt: { type: string }
    Role:
      type: object
      required: [id, name, code, description]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        description: { type: string }
        dataScope: { type: string }
        permissions: { type: array, items: { $ref: "#/components/schemas/Permission" } }
    Permission:
      type: object
      required: [id, name, code, path, method, kind, description]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        path: { type: string }
        method: { type: string }
        kind: { type: string }
        description: { type: string }
        parentId: { type: integer, nullable: true }
        sort: { type: integer }
        icon: { type: string }
        routePath: { type: string }
        component: { type: string }
        hidden: { type: boolean }
...
    OpLog:
      type: object
      required: [id, username, module, action, method, path, status, ip, latencyMs, createdAt]
      properties:
        id: { type: integer }
        traceId: { type: string }
        username: { type: string }
        module: { type: string }
        action: { type: string }
        method: { type: string }
        path: { type: string }
        status: { type: integer }
        ip: { type: string }
        latencyMs: { type: integer }
        detail: { type: string }
        description: { type: string }
        oldValue: { type: string }
        newValue: { type: string }
        createdAt: { type: string }
...
    Department:
      type: object
      required: [id, name, code, sort, leader, status]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        parentId: { type: integer, nullable: true }
        sort: { type: integer }
        leader: { type: string }
        status: { type: string }
        children: { type: array, items: { $ref: "#/components/schemas/Department" } }
    DictType:
      type: object
      required: [id, code, name, status, remark]
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        status: { type: string }
        remark: { type: string }
...
    SysConfig:
      type: object
      required: [id, key, value, name, group, remark]
      properties:
        id: { type: integer }
        key: { type: string }
        value: { type: string }
        name: { type: string }
        group: { type: string }
        remark: { type: string }
...
    PageMeta:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    EnvelopeUserPage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/User" } }
    EnvelopeOpLogPage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/OpLog" } }
...
    EnvelopeDepartmentPage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/Department" } }
    EnvelopeRolePage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/Role" } }
    EnvelopePermissionPage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/Permission" } }
    EnvelopeDictPage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/DictType" } }
...
    EnvelopeConfigPage:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageMeta"
                - type: object
                  properties:
                    items: { type: array, items: { $ref: "#/components/schemas/SysConfig" } }
...